Neural network now works inside of UE, generates a prediction from dummy data.

+++ TFInUnreal/Source/TFInUnreal/TensorflowNetwork.h
@@ -79,5 +79,5 @@ private:
 	UMaterialInstanceDynamic* DynamicMaterial;
 
 	bool Puk{ false };
-	std::vector<float> Result;
+	std::vector<double> Result;
 };
